Move compile to AstExpression, remove Operation, and update all associated calls.
